✅ Refactor test with explicit stringified annotations
